Updated Tailwind config, Extracted form to component
