+++ client/tailwind.config.js
@@ -5,5 +5,15 @@ module.exports = {
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
   ],
+  theme: {
+    extend: {
+      colors: {
+        success:'#00ff00',
+        error: '#ff0000',
+        warning: '#ffff00',
+        info: '#0000ff',
+      },
+    },
+  },
   plugins: [],
 }
